added explanatory comments and deleted exploratory code snippets
