+++ server.py
@@ -1,6 +1,8 @@
 import http.server
 import socketserver
 
+#Sets up a simple HTTP server
+
 PORT = 8000
 Handler = http.server.SimpleHTTPRequestHandler
